feat: auto-close agents
