+++ codex-rs/core/src/tools/handlers/agent_jobs.rs
@@ -248,7 +248,7 @@ async fn run_agent_job_loop(
                     let _ = session
                         .services
                         .agent_control
-                        .shutdown_live_agent(thread_id)
+                        .shutdown_and_forget_agent(thread_id)
                         .await;
                     continue;
                 }
@@ -369,7 +369,7 @@ async fn recover_running_items(
                 let _ = session
                     .services
                     .agent_control
-                    .shutdown_live_agent(thread_id)
+                    .shutdown_and_forget_agent(thread_id)
                     .await;
             }
             continue;
@@ -491,7 +491,7 @@ async fn reap_stale_active_items(
         let _ = session
             .services
             .agent_control
-            .shutdown_live_agent(thread_id)
+            .shutdown_and_forget_agent(thread_id)
             .await;
         active_items.remove(&thread_id);
     }
@@ -527,7 +527,7 @@ async fn finalize_finished_item(
     let _ = session
         .services
         .agent_control
-        .shutdown_live_agent(thread_id)
+        .shutdown_and_forget_agent(thread_id)
         .await;
     Ok(())
 }
